Refactoring and bug fixes in the store.
Tried to work within the confines of the linkbench framework which doesn't really fit the blueprints model, leaving the implementation a bit sub-optimal overall.  Larger changes would be needed to make this really be more friendly to plugging in new data stores to LinkBench, but none are "simple" changes at the moment.

Changed methods that get links to return null when source vertices are not found.  That seems to be model in the MySQL implementation if I've interpreted it correctly.